Move main entry point to src directory; update paths in package.json and related files
