+++ package.json
@@ -3,11 +3,11 @@
   "version": "0.0.3",
   "description": "A tool to validate Markdown and AsciiDoc structure against a specified template",
   "bin": {
-    "doc-structure-lint": "index.js"
+    "doc-structure-lint": "src/index.js"
   },
-  "main": "index.js",
+  "main": "src/index.js",
   "scripts": {
-    "start": "node index.js",
+    "start": "node src/index.js",
     "test": "mocha 'src/**/*.test.js'",
     "prepare": "husky"
   },
